requirements/base.txt

# Django
# ------------------------------------------------------------------------------
django==4.1 # https://www.djangoproject.com/
psycopg2==2.9.4
environ==1.0


django-filter==2.0.0
drf-spectacular==0.24.2

